localizse/requirements.txt
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
autopep8==1.5.7
black==21.11b1
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
entrypoints==0.3
flake8==4.0.1
git-filter-repo==2.34.0
google-api-core==1.31.4
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.54.0
grpcio==1.42.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jellyfish==0.8.8
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lib50==3.0.3
markdown2==2.4.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==8.3.2
platformdirs==2.4.0
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-debian===0.1.36ubuntu1
pytz==2021.3
PyYAML==5.3.1
regex==2021.9.30
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
submit50==3.1.0
systemd-python==234
termcolor==1.1.0
toml==0.10.2
tomli==1.2.2
Twisted==18.9.0
typing-extensions==4.0.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
